requirements.txt
fastapi[all]
uvicorn
google-api-python-client
SQLAlchemy
pydantic
python-dotenv
httpx
